rename a misspelling
