+++ src/server.ts
@@ -1,6 +1,6 @@
 import app from "./app.ts";
 import { PORT } from "./config/env.ts";
-import { connectDB } from "./config/mongobd.ts";
+import { connectDB } from "./config/mongodb.ts";
 
 app.listen(PORT, async () => {
   console.log(`🚀 Server running at http://localhost:${PORT}`);
